Various changes.
