--- .ddev/commands/web/test-performance
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-## Description: Run performance tests
-## Usage: test-performance
-
-set -e
-php scale.php
