+++ .gitignore
@@ -52,9 +52,6 @@ node_modules/
 # Current
 /.dumps/cms-prod-files-latest.tgz
 /.dumps/cms-prod-db-sanitized-latest.sql*
-# Deprecated
-/.dumps/cms-app-files-latest.tgz
-/.dumps/cms-db-latest.sql*
 
 # Ignore common IDE configs
 # IntelliJ/PhpStorm
